bh_asciitable - generate the table on the fly
It's fine to have a static table if you want to, but I think it's neater to
generate it on demand instead. My generator has exacly preserved the output
from your static version even in places where doing so was inconvenient, and
even in places where I think the results are not as good as doing something
more regular.

If you were to have a static table it would be better to generate it without
employing echo, and especially without using echo with switches e.g. -en.

The echo command should be, IMNSHO, treated as historical. If you want a
command which will, as with echo -e, expand escape sequences, and will, as
with echo -n, omit a trailing newline, then what you want is not echo at all
but printf, which is the command that replaced echo in 1986 (or, if you are
conservative, perhaps as late as 1990).

See https://www.in-ulm.de/~mascheck/various/echo+printf/ for a good rundown on
the situation with echo and printf. In brief, echo with switches is not safely
portable and, what's more, the standards explicitly require switches to not be
supported. In my own personal shell I redefine echo with this function:
	echo () { builtin printf -- '%s\n' "$*"; }
so uses which presume switches will work stand out to me. I strongly urge you
not to employ echo at all, or at a minimum to do so only without any switches.