bh_hex2str: avoid external utilities
A pure-bash rewrite that produces identical output. I have also added support
for taking in more than a single argument; if more are passed each will be
emitted on its own line.

I have also gone out of my way to avoid the use of echo. IMNSHO echo should
be considered historical and should not be used, especially not with any
switches.