bh_ip2bin: a pure bash implementation
The use of echo here was unnecessary and inefficient.

Consider the difference between this:

	echo 1.2.3.4 | tr . ' '

and this:

	tr . ' ' <<<1.2.3.4

As far as tr is concerned they are the same, but the latter is more efficient
because an additional subshell is not created for the pipe.

But, there is no need to employ tr at all. What you had been doing was
creating whitespace and then relying on the shell to do word splitting later
to create an argument list for bh_dec2bin, but you can split the IP string
with read instead and generate an array which can then be expanded directly
into arguments to bh_dec2bin without abusing shell semantics.

Since an aray was created anyway it is then easy to avoid the need to print
and then trim a trailing dot. You could always have arranged to avoid the use
of sed by simply appending onto a shell variable and the trimming it in the
end, e.g.:

	bin+=$(printf %.8d. ...)
	# ... later
	bin=${bin%.}

... but with each octet stored separately in a list this is even easier.