bh_hex2bin: a pure bash version
The original version iterated over an UNQUOTED $*! You should never rely on
the behavior of unquoted expansion unless you are constrained to POSIX sh and
have literally no other option. There are other, better options in this case;
even if POSIX sh you can iterate over quoted $@ to iterate over each argument
to the function. The only observable difference would be what happens if you
pass in a *quoted* argument with whitespace in it, e.g. in this version:

	$ bh_hex2bin '1 2 3'
	-bash ((: 16#1 2 3 > 0: syntax error in expression (error token is "2 3 > 0")

But in the original:

	$ bh_hex2bin '1 2 3'
	1 10 11

If you really wanted to support that you could accomplish it by processing
each input argument through read, e.g.

	IFS=' ' read -ra n_list <<<"$n"

And then iterate over n_list. Although this is more lines of code the result
is safer and more reliable than the original implementation. I chose to treat
such input as a bug instead; arguably I should check for invalid input and
abort with a nicer error message, but I think the existing error is clear
enough.

The original version also performed unquoted expansion of $i in a location
where doing so is unsafe. You should not expand any variables unquoted in
bash; if you think you need to you have a bug and should find another
solution.

I have also replaced all uses of echo with printf. IMNSHO, echo should be
treated as deprecated and should not be used for any reason.