str2hex: stylistic cleanup
A variable expanded as a parameter to "case" is implicitly expanded as if
double-quoted. Although the double quotes do no harm, and when in doubt you
should always prefer to double quote expansions, in this scenario it safe to
leave it unquoted and I have made that change here/

I have also un-quoted the pattern values in the case statement. Quoting them
does make them literal, but it is also not required when meta characters are
not in use, which they are not in this case.

There was some inconsistent indentation on the default cause which has been
normalized to match the other cases.

Finally, I have moved the position of the ;; that closes a case condition to
align with the match pattern. I prefer this style as to me it forms a kind o
a block. I consider this:

	case $1 in
		foo)
			printf '%s\n' "$1"
		;;
	esac

To be equivalent to this C-style switch:

	switch (argv[1]) {
		case "foo": {
			printf("%s\n", argv[1]);
			break;
		}
	}

Thus, the close ;; is equivalent to the close curly brace at the end of the
"case" block.