bh_bin2ip: always use -r with read
Without -r the read command probably isn't doing what you think it's doing.
For historical reasons it will interpret backslash in the input as the start
of an escape sequence; understood escapes will be replaced, but even if the
next character isn't a known escape char the \ will be removed. This is
almost never what you *intended* to ask for.

I have added -r to read here, and unless you *intend* backslash removal like
that you should do that everywhere.