bh_hex2dec: avoid the use of echo, minor cleanup
I've changed the guard check at the beginning to one that is slightly more
concise. In a numeric context a zero evaluates to a return code of true, so
simply evaluating the argument count will automatically give you true/false
for zero/non-zero args.

I have replaced the use of echo with printf. IMNSHO echo should be considered
deprecated and should not be used for any reason.

I have used numeric context base notation to force the value of $1 to be
treated as hex just because I think it's tidier.

Finally, I've replaced space indent with tab indent. Most of the lines in
most files are already indented with tab, and I believe it is better in
general.